# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "EvidenceResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "document_name": {
            "title": "Document Name",
            "type": "string"
          },
          "excerpt": {
            "minLength": 1,
            "title": "Excerpt",
            "type": "string"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "page_number": {
            "minimum": 1.0,
            "title": "Page Number",
            "type": "integer"
          },
          "section_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Name"
          }
        },
        "required": [
          "document_id",
          "document_name",
          "page_number",
          "excerpt",
          "id",
          "finding_id",
          "created_at"
        ],
        "title": "EvidenceResponse",
        "type": "object"
      },
      "FindingCategory": {
        "description": "Which screening dimension a finding belongs to.\n\nThe first four mirror the four scoring categories of spec \u00a714.\n``DATA_COMPLETENESS`` carries known-unknowns that carry no points.",
        "enum": [
          "site_suitability",
          "environmental",
          "access",
          "permitting",
          "data_completeness"
        ],
        "title": "FindingCategory",
        "type": "string"
      },
      "FindingGroup": {
        "description": "How a finding should be presented (spec \u00a710.4).\n\nMissing information is kept separate from confirmed risks: an absent input is\nnot evidence of a problem, and the two must never be conflated in the UI.\n\n``REQUIREMENT`` was added at integration for document-derived permitting\nrequirements. Filing them under ``RISK`` would assert that \"a conditional-use\npermit is required\" is a problem with the site, which it is not \u2014 it is an\nobligation, and it gets its own list on the site page.",
        "enum": [
          "positive_signal",
          "risk",
          "missing_information",
          "requirement"
        ],
        "title": "FindingGroup",
        "type": "string"
      },
      "FindingSeverity": {
        "description": "Spec \u00a713. A single ``FATAL`` finding forces a ``REJECT`` recommendation.",
        "enum": [
          "info",
          "warning",
          "high",
          "fatal"
        ],
        "title": "FindingSeverity",
        "type": "string"
      },
      "FindingSourceType": {
        "description": "Spec \u00a715.5 \u2014 how a finding was produced.",
        "enum": [
          "deterministic",
          "document",
          "human"
        ],
        "title": "FindingSourceType",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "minimum_acres": {
            "title": "Minimum Acres",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "project_type": {
            "$ref": "#/components/schemas/ProjectType"
          },
          "screening_criteria": {
            "$ref": "#/components/schemas/ScreeningCriteria"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "target_capacity_mw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Capacity Mw"
          },
          "target_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target State"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "project_type",
          "target_capacity_mw",
          "minimum_acres",
          "target_state",
          "screening_criteria",
          "notes",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectRead",
        "type": "object"
      },
      "ProjectStatus": {
        "description": "Serialized as a plain string, which is all the v1 contract promises.",
        "enum": [
          "active",
          "screening",
          "screened"
        ],
        "title": "ProjectStatus",
        "type": "string"
      },
      "ProjectType": {
        "description": "Spec \u00a710.2. ``COMMUNITY_SOLAR`` comes from the \u00a716.1 request example.",
        "enum": [
          "solar",
          "community_solar",
          "battery_storage",
          "data_center",
          "ev_charging",
          "other"
        ],
        "title": "ProjectType",
        "type": "string"
      },
...
      "RequirementCategory": {
        "description": "The kind of permitting obligation a document-derived finding states.\n\nDistinct from ``FindingCategory``, which says which of the four scoring\ndimensions a finding belongs to. A document finding is always\n``FindingCategory.PERMITTING``; this says what sort of permitting item it is.",
        "enum": [
          "use_permission",
          "setback",
          "public_hearing",
          "decommissioning",
          "financial_security",
          "environmental_study",
          "traffic_study",
          "application_requirement",
          "other"
        ],
        "title": "RequirementCategory",
        "type": "string"
      },
      "ReviewDecision": {
        "description": "A reviewer's action on a finding. Maps onto ``ReviewStatus``.",
        "enum": [
          "approve",
          "edit",
          "reject",
          "escalate"
        ],
        "title": "ReviewDecision",
        "type": "string"
      },
      "ReviewFindingRequest": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/ReviewDecision"
          },
          "edited_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Description"
          },
          "edited_title": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Title"
          },
          "reviewer_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Note"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ReviewFindingRequest",
        "type": "object"
      },
      "ReviewStatus": {
        "description": "State of a finding in the human-review loop (spec \u00a78.1F).",
        "enum": [
          "pending",
          "approved",
          "edited",
          "rejected",
          "escalated"
        ],
        "title": "ReviewStatus",
        "type": "string"
      },
      "RiskFindingRead": {
        "description": "A persisted finding as returned by the API, whatever produced it.",
        "properties": {
          "actual_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Value"
          },
          "category": {
            "$ref": "#/components/schemas/FindingCategory"
          },
          "confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          },
          "group": {
            "$ref": "#/components/schemas/FindingGroup"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "original_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Description"
          },
          "original_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Title"
          },
          "requirement_category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RequirementCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "requires_human_review": {
            "title": "Requires Human Review",
            "type": "boolean"
          },
          "review_status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "rule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule"
          },
          "screening_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screening Run Id"
          },
          "severity": {
            "$ref": "#/components/schemas/FindingSeverity"
          },
          "site_id": {
            "title": "Site Id",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/FindingSourceType"
          },
          "threshold_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold Value"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "id",
          "site_id",
          "screening_run_id",
          "source_type",
          "category",
          "group",
          "rule",
          "title",
          "description",
          "severity",
          "value",
          "actual_value",
          "threshold_value",
          "confidence",
          "review_status",
          "requirement_category",
          "original_title",
          "original_description",
          "requires_human_review",
          "evidence",
          "created_at",
          "updated_at"
        ],
        "title": "RiskFindingRead",
        "type": "object"
      },
...
      "ScreeningRunDetail": {
        "description": "A run plus its ranked results, highest score first.\n\nThe project is embedded because every screen that shows a ranking also shows\nthe criteria the ranking was produced against; sending them together means the\nUI cannot render a score next to a stale threshold.",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectRead"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SiteScreeningResult"
            },
            "title": "Results",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/ScreeningRunStatus"
          }
        },
        "required": [
          "id",
          "project_id",
          "status",
          "idempotency_key",
          "started_at",
          "completed_at",
          "error_message",
          "created_at",
          "project",
          "results"
        ],
        "title": "ScreeningRunDetail",
        "type": "object"
      },
      "ScreeningRunStatus": {
        "description": "Spec \u00a718 \u2014 partial completion is a first-class outcome.",
        "enum": [
          "queued",
          "screening",
          "document_analysis",
          "needs_review",
          "completed",
          "partially_completed",
          "failed"
        ],
        "title": "ScreeningRunStatus",
        "type": "string"
      },
...
      "SiteScreeningResult": {
        "description": "One ranked site, with everything needed to explain its result.\n\nThe counts and the next action are computed here rather than in the UI. They\nare conclusions drawn from findings and a recommendation status, and the\nbackend is the only place allowed to draw them (CLAUDE.md, spec \u00a79.1).",
        "properties": {
          "high_risk_finding_count": {
            "title": "High Risk Finding Count",
            "type": "integer"
          },
          "missing_information": {
            "items": {
              "$ref": "#/components/schemas/RiskFindingRead"
            },
            "title": "Missing Information",
            "type": "array"
          },
          "permitting_requirements": {
            "items": {
              "$ref": "#/components/schemas/RiskFindingRead"
            },
            "title": "Permitting Requirements",
            "type": "array"
          },
          "positive_signals": {
            "items": {
              "$ref": "#/components/schemas/RiskFindingRead"
            },
            "title": "Positive Signals",
            "type": "array"
          },
          "recommended_next_action": {
            "title": "Recommended Next Action",
            "type": "string"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/RiskFindingRead"
            },
            "title": "Risks",
            "type": "array"
          },
          "score": {
            "$ref": "#/components/schemas/SiteScoreRead"
          },
          "site": {
            "$ref": "#/components/schemas/CandidateSiteRead"
          },
          "warning_count": {
            "title": "Warning Count",
            "type": "integer"
          }
        },
        "required": [
          "site",
          "score",
          "positive_signals",
          "risks",
          "permitting_requirements",
          "missing_information",
          "high_risk_finding_count",
          "warning_count",
          "recommended_next_action"
        ],
        "title": "SiteScreeningResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowEventRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Summary"
          },
          "output_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Summary"
          },
          "screening_run_id": {
            "title": "Screening Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "step_name": {
            "title": "Step Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "screening_run_id",
          "step_name",
          "status",
          "input_summary",
          "output_summary",
          "duration_ms",
          "error_message",
          "created_at"
        ],
        "title": "WorkflowEventRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Evidence-backed site triage for renewable energy and power infrastructure projects. Foundation build \u2014 endpoints are added by feature agents (see docs/PARALLEL_TASKS.md).",
    "title": "SiteSift API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/findings/{finding_id}/review": {
      "patch": {
        "description": "Approve, edit, reject, or escalate a finding.\n\nAppend-only: each decision writes a new ``reviews`` row and the extracted\noriginal is preserved, so an edit never destroys what the document actually\nsaid (spec \u00a79.3).",
        "operationId": "review_site_finding_api_findings__finding_id__review_patch",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewFindingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskFindingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Site Finding",
        "tags": [
          "findings"
        ]
      }
    },
...
    "/api/projects/{project_id}/screenings": {
      "post": {
        "description": "Screen every candidate site in the project.\n\nSpec \u00a718: a repeated request carrying the same ``Idempotency-Key`` returns\nthe run that already exists \u2014 200, not 201 \u2014 rather than screening twice.",
        "operationId": "create_screening_api_projects__project_id__screenings_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreeningRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Screening",
        "tags": [
          "screenings"
        ]
      }
    },
    "/api/projects/{project_id}/screenings/latest": {
      "get": {
        "description": "The project's most recent run, with its ranked results.\n\nThe results screen needs the project, the run, and the ranking together. 404\nmeans the project has not been screened yet \u2014 a normal state the UI shows as\nan empty state, not an error.",
        "operationId": "get_latest_screening_api_projects__project_id__screenings_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreeningRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Latest Screening",
        "tags": [
          "screenings"
        ]
      }
    },
...
    "/api/screenings/{screening_id}": {
      "get": {
        "operationId": "get_screening_api_screenings__screening_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "screening_id",
            "required": true,
            "schema": {
              "title": "Screening Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreeningRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Screening",
        "tags": [
          "screenings"
        ]
      }
    },
    "/api/screenings/{screening_id}/events": {
      "get": {
        "description": "The audit trail for a run, in the order the steps ran (spec \u00a718).",
        "operationId": "list_screening_events_api_screenings__screening_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "screening_id",
            "required": true,
            "schema": {
              "title": "Screening Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowEventRead"
                  },
                  "title": "Response List Screening Events Api Screenings  Screening Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Screening Events",
        "tags": [
          "screenings"
        ]
      }
    },
...
    "/api/sites/{site_id}/findings": {
      "get": {
        "description": "Every finding for a site \u2014 deterministic and document-derived alike.\n\nEvidence is eager-loaded: a document-derived claim is only meaningful with the\nexcerpt it rests on, and the UI must never show one without the other.",
        "operationId": "list_site_findings_api_sites__site_id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RiskFindingRead"
                  },
                  "title": "Response List Site Findings Api Sites  Site Id  Findings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Site Findings",
        "tags": [
          "findings"
        ]
      }
    },
...
  }
}